requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1623861369333/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1632748415174/work
astor==0.8.1
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
blinker==1.4
brotlipy==0.7.0
bsddb3==6.2.6
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2022.5.18.1
cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018138503/work
configparser @ file:///tmp/build/80754af9/configparser_1614614799251/work
cryptography @ file:///tmp/build/80754af9/cryptography_1607540692013/work
cycler==0.10.0
Cython==0.29.30
cytoolz==0.11.0
dask==2021.10.0
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
easydict==1.9
faiss==1.7.0
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fsspec @ file:///opt/conda/conda-bld/fsspec_1652802105984/work
future==0.18.2
gast==0.2.2
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1597424467487/work
h5py==2.8.0
hdbscan==0.8.28
huggingface-hub==0.4.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1631906757572/work
ipython @ file:///tmp/build/80754af9/ipython_1632141056704/work
jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
lmdb==1.3.0
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
Markdown @ file:///tmp/build/80754af9/markdown_1614363832606/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846017414/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mrcfile==1.3.0
multidict @ file:///tmp/build/80754af9/multidict_1607367738239/work
munkres==1.1.4
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
nose @ file:///tmp/build/80754af9/nose_1606773131901/work
numpy==1.21.6
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.64
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.2.4
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pkginfo==1.7.0
prettytable==3.2.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.17.2
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==7.0.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyefd==1.6.0
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
PyJWT==1.7.1
PyOpenGL==3.1.5
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658308664/work
PyYAML==5.4.1
regex==2022.3.15
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ruamel-yaml==0.15.87
sacremoses==0.0.49
scikit-image==0.18.3
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370326158/work
scipy @ file:///tmp/build/80754af9/scipy_1616700973326/work
six==1.15.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1614023247537/work
sphire @ file:///home/eman2/miniconda3/conda-bld/eman2_1615224802990/work/sphire
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2020.10.1
tokenizers==0.11.6
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchsummary==1.5.1
torchtoolbox==0.1.8.2
torchvision==0.11.1+cu113
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm @ file:///tmp/build/80754af9/tqdm_1611857934208/work
traitlets @ file:///tmp/build/80754af9/traitlets_1632522747050/work
transformers==4.17.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1631818832573/work
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
